requirements.txt
py-solc-x
slither-analyzer
